#23 chore. 제목 줄바꿈 표시 문제 수정
