+++ app/src/main/java/com/bonda/bonda/ui/profile/recent/articles/ArticleAdapter.kt
@@ -34,7 +34,7 @@ class ArticleAdapter(
         private val categoryChip: Chip = view.findViewById(R.id.category)
 
         fun bind(article: Article) {
-            titleTv.text      = article.title
+            titleTv.text      = article.title.replace("\\n", "\n")
             imageView.load(article.imageUrl)
             categoryChip.text = article.category
