+++ app/src/main/java/com/bonda/bonda/ui/search/SearchResultFragment.kt
@@ -18,6 +18,7 @@ import com.bonda.bonda.databinding.FragmentSearchResultAllBinding
 import com.bonda.bonda.databinding.ViewArticleBinding
 import com.bonda.bonda.databinding.ViewBookVerticalBinding
 import com.bonda.bonda.databinding.ViewRecentArticleBinding
+import com.bonda.bonda.model.toArticleCategory
 import com.bonda.bonda.ui.article.ArticleActivity
 import com.bonda.bonda.ui.book.BookActivity
 import com.bonda.bonda.util.TAG
@@ -128,8 +129,8 @@ class SearchResultFragment : Fragment() {
                     false
                 )
 
-                itemBinding.title.text = article.title
-                itemBinding.category.root.text = article.category
+                itemBinding.title.text = article.title.replace("\\n", "\n")
+                itemBinding.category.root.text = article.category.toArticleCategory().label
                 itemBinding.image.load(article.imageUrl)
 
                 val params = itemBinding.root.layoutParams as GridLayout.LayoutParams
